NEW Support for other authenticators
I've added the ability to add other authenticators to the RestfulServer

Use the config system to change the authenticator option and it'll call
'authenticate' on that class.

I've assumed that one won't need any parameters sent to that and
everything can be taken from `Controller::curr()->request`

This relies on the default authenticator being able to accept the
standard email/password identifiers. I imagine anyone using a default
authenticator to that will go to the trouble of adding a RestfulServer
authenticator too.